Add the missing documentation
