--- src/AdventOfCode.Commons/Input/IPuzzleInput.cs
@@ -1,6 +1,0 @@
-﻿namespace AdventOfCode.Commons.Input;
-
-public interface IPuzzleInputReaderStrategy
-{
-    IEnumerable<string> ReadInput();
-}
