--- src/AdventOfCode.Commons/Input/LocalPuzzleInputReaderStrategy.cs
@@ -1,9 +1,0 @@
-﻿namespace AdventOfCode.Commons.Input;
-
-public class LocalPuzzleInputReaderStrategy : IPuzzleInputReaderStrategy
-{
-    public required string InputPath { get; init; }
-
-    public IEnumerable<string> ReadInput()
-        => File.ReadAllLines(InputPath);
-}
